ml-service/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
numpy
pymongo
python-dotenv
